requirements.txt
gtts
streamlit
python-dotenv
langchain_google_genai
langchain
PyPDF2


-e .